0604 from h

+++ XunLinMineRemoteControlWeb/XunLinMineRemoteControlWeb/Core/WebLoginUserInfo.cs
@@ -15,7 +15,7 @@ namespace XunLinMineRemoteControlWeb.Core
 
         public string Token;
 
-        public int ShoppingCredits;
+        public decimal ShoppingCredits;
 
         public string UserRemoteServerValidStopTimeText = "";
